Correccion de entities

+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/CondicionesEspecialesVehiculoDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class CondicionesEspecialesVehiculoDao extends GenericDao<CondicionesEspecialesVehiculo, Long> {
+public class CondicionesEspecialesVehiculoDao extends GenericDao<CondicionesEspecialesVehiculo, Integer> {
 
 	
 	public CondicionesEspecialesVehiculoDao(EntityManager entityManager) {
