+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/CoordenadaDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class CoordenadaDao extends GenericDao<Coordenada, Long> {
+public class CoordenadaDao extends GenericDao<Coordenada, Integer> {
 
 	
 	public CoordenadaDao(EntityManager entityManager) {
