+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/DireccionDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class DireccionDao extends GenericDao<Direccion, Long> {
+public class DireccionDao extends GenericDao<Direccion, Integer> {
 
 	
 	public DireccionDao(EntityManager entityManager) {
