+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/EncomiendaDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class EncomiendaDao extends GenericDao<Encomienda, Long> {
+public class EncomiendaDao extends GenericDao<Encomienda, Integer> {
 
 	
 	public EncomiendaDao(EntityManager entityManager) {
