+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/RolUsuarioDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class RolUsuarioDao extends GenericDao<RolUsuario, Long> {
+public class RolUsuarioDao extends GenericDao<RolUsuario, Integer> {
 
 	
 	public RolUsuarioDao(EntityManager entityManager) {
