+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/SucursalDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class SucursalDao extends GenericDao<Sucursal, Long> {
+public class SucursalDao extends GenericDao<Sucursal, Integer> {
 
 	
 	public SucursalDao(EntityManager entityManager) {
