+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/TareaMantenimientoRealizadaDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class TareaMantenimientoRealizadaDao extends GenericDao<TareaMantenimientoRealizada, Long> {
+public class TareaMantenimientoRealizadaDao extends GenericDao<TareaMantenimientoRealizada, Integer> {
 
 	
 	public TareaMantenimientoRealizadaDao(EntityManager entityManager) {
