+++ 2016AD_G04_Servidor/src/main/java/com/ADG04/Servidor/dao/UsuarioDao.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-public class UsuarioDao extends GenericDao<Usuario, Long> {
+public class UsuarioDao extends GenericDao<Usuario, Integer> {
 
 	
 	public UsuarioDao(EntityManager entityManager) {
